// src/modules/user/dto/returnUser.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { Expose } from 'class-transformer';

export default class ReturnUserDto {
  @ApiProperty({
    example: '550e8400-e29b-41d4-a716-446655440000',
    description: 'unique uuid generated by db',
  })
  @Expose()
  uuid: string;

  @ApiProperty({
    example: 'user@example.com',
    description: 'Unique email of a user',
  })
  @Expose()
  email: string;
}